Tied in the pages for renters and landlords to enter their intitial information. TODO: prevent required fields from being entered empty
