+++ pages/landlord/walkthrough/2.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { UserApi } from "../../../generated-src/openapi";
 import { useRouter } from "next/router";
-import { Error, Header, Button, ProgressBar, CenteredImage, PulseLottie } from "../../../components";
+import { Error, Header, Button, ProgressBar, CenteredImage } from "../../../components";
 
 export default function RenterWalkthrough1() {
     const router = useRouter();
@@ -24,7 +24,6 @@ export default function RenterWalkthrough1() {
             });
         return () => sub.unsubscribe();
     }, [router.isReady, router.query]);
-
     return (
         <div>
             <Header
@@ -34,14 +33,17 @@ export default function RenterWalkthrough1() {
                 showBack={false}
                 showLogout={false}
             />
-            <ProgressBar progress="2/6" />
+            <ProgressBar progress="2/5" />
             <div className="place-items-center">
-                <p className="text-center tk-text-teal text-3xl font-semibold pt-5">Next...</p>
-                <p className="text-center tk-text-blue text-medium text-medium pt-8 px-8 h-32">Search for listings
-                    on your preferred listing platform.</p>
-                <div className="static h-96"> 
-                    <CenteredImage src="/assets/images/renter_profile.png" alt="renter profile" className="w-screen h-full"></CenteredImage>
-                    <PulseLottie left={200} top={525}  width={100} height={100}/>
+                <p className="text-center tk-text-teal text-3xl font-semibold pt-5">
+                    Next...
+                </p>
+                <p className="text-center tk-text-blue text-medium pt-8 px-8 h-32">
+                    A renter sends you a link to view their Turnkey Renter profile.
+                </p>
+                <div className="flex flex-col space-y-4 h-96"> 
+                    <CenteredImage src="/assets/images/rental_url_share.png" className="w-screen h-full"></CenteredImage>
+                    <CenteredImage src="/assets/images/default_user_icon.png" className="w-screen h-full"></CenteredImage>
                 </div>
                 <div className="flex flex-col gap-5 px-16 absolute w-screen bottom-4">
                     <Button variant="secondary" handleClick={() =>
