+++ pages/landlord/walkthrough/3.tsx
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from "react";
-import { UserApi } from "../../../generated-src/openapi";
+import { UpdateUserRequest, UserApi, UserDto } from "../../../generated-src/openapi";
+import { Observable } from "rxjs";
 import { useRouter } from "next/router";
-import { Error, Header, Button, ProgressBar, CenteredImage, PulseLottie} from "../../../components";
+import { Header, Button, ProgressBar, CenteredImage, PulseLottie } from "../../../components";
 
 export default function RenterWalkthrough1() {
     const router = useRouter();
@@ -24,6 +25,30 @@ export default function RenterWalkthrough1() {
             });
         return () => sub.unsubscribe();
     }, [router.isReady, router.query]);
+
+    function completeWalkthrough() {
+        const userApi = new UserApi();
+        user.walkthroughComplete = true;
+
+        let obs: Observable<void>;
+        let body = user as UserDto;
+
+        const updateRequest = {
+            id: router.query.userId as string,
+            token: router.query.token as string,
+            body: body
+        } as UpdateUserRequest;
+        
+        obs = userApi.updateUser(updateRequest);
+
+        obs.subscribe(() =>
+            router.push({
+                pathname: "/landlord",
+                query: { userId: router.query.userId, token: router.query.token },
+            })
+        );
+    }
+
     return (
         <div>
             <Header
@@ -33,17 +58,15 @@ export default function RenterWalkthrough1() {
                 showBack={false}
                 showLogout={false}
             />
-            <ProgressBar progress="3/6" />
+            <ProgressBar progress="3/5" />
             <div className="place-items-center">
-                <p className="text-center tk-text-teal text-3xl font-semibold pt-5">
-                    After you find a place you like...
-                </p>
+                <p className="text-center tk-text-teal text-3xl font-semibold pt-5">View your Applicants' Profile</p>
                 <p className="text-center tk-text-blue text-medium pt-8 px-8 h-32">
-                    Share your Turnkey profile by copying your unique Renter ID link.
+                    A renter submits their profile to you which can be used as a rental application.
                 </p>
                 <div className="static h-96"> 
-                    <CenteredImage className="w-screen h-full" src="/assets/images/share_renter_profile.png" alt="renter profile"></CenteredImage>
-                    <PulseLottie left={200} top={525} width={50} height={50}/>
+                    <CenteredImage className="w-screen h-full" src="/assets/images/renter_add.png" alt="renter profile" ></CenteredImage>
+                    <PulseLottie left={200} top={450} width={50} height={50}/>
                 </div>
                 <div className="flex flex-col gap-5 px-16 absolute w-screen bottom-4">
                     <Button variant="secondary" handleClick={() =>
