+++ pages/landlord/walkthrough/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { UserApi } from "../../../generated-src/openapi";
 import { useRouter } from "next/router";
-import { Error, Header, Button, ProgressBar, CenteredImage, PulseLottie } from "../../../components";
+import { Header, Button, ProgressBar, CenteredImage, PulseLottie } from "../../../components";
 
 export default function RenterWalkthrough1() {
     const router = useRouter();
@@ -33,15 +33,14 @@ export default function RenterWalkthrough1() {
                 showBack={false}
                 showLogout={false}
             />
-            <ProgressBar progress="1/6" />
+            <ProgressBar progress="1/5" />
             <div className="place-items-center">
                 <p className="text-center tk-text-teal text-3xl font-semibold pt-5">First...</p>
-                <p className="text-center tk-text-blue text-medium pt-8 px-8 h-32">Complete your
-                    Turnkey profile with all the information
-                    required on a standard rental application.</p>
-                <div className="static h-96"> 
-                    <CenteredImage className="w-screen h-full" src="/assets/images/renter_profile.png" alt="renter profile" ></CenteredImage>
-                    <PulseLottie left={200} top={400} width={100} height={100}/>
+                <p className="text-center tk-text-blue text-medium pt-8 px-8 h-32">
+                    A renter finds your property to rent on any listing platform.
+                </p>
+                <div className="static h-96">
+                    <CenteredImage className="w-screen h-full" src="/assets/images/rental_sites.png" alt="renter profile" ></CenteredImage>
                 </div>
                 <div className="flex flex-col gap-5 px-16 absolute w-screen bottom-4">
                     <Button variant="secondary" handleClick={() =>
